pyproject.toml
[tool.poetry]
name = "reckoning"
version = "0.1.0"
description = ""
authors = ["Ruslan Sayfullin <ruslansaifullin91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-debug-toolbar = "^3.8.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.2"
drf-nested-routers = "^0.93.4"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
